Find positions of number in array - Binary search
There are a sorted array (non-decreasing order) and a target number. The task is to find positions where this target is in the array and return a list [first_position, last_position]. If this number is not in the array - return [-1, -1].
The solution here is with O(log n) runtime complexity. I used binary search three times: to find any position of the number, then to find the first one and the last one.